Protest data

Protest data

Usage

protest

Format

protest

A data frame with 6 variables, 129 cases:

subnum

subject number

protest

experimental condition, 0 = no protest, 1 = individual protest, 2 = group protest

sexism

perceived pervasiveness of sex discrimination. Means of an 8 item Modern Sexism Scale

angry

anger toward the attorney. "I feel angry towards Catherine"

liking

liking of the attorney. Mean rating of 6 liking ratings of the target

respappr

appropriateness of response. Mean of four items of appropriateness of the target's response

Source

Garcia, D. M., Schmitt, M. T., Branscombe, N. R., & Ellemers, N. (2010). Women’s reactions to ingroup members who protest discriminatory treatment: The importance of beliefs about inequality and response appropriateness. European Journal of Social Psychology, 40(5), 733–745. https://doi.org/10.1002/ejsp.644

Hayes, A. (2017). Introduction to mediation, moderation, and conditional process analysis: A regression-based approach (2nd edn, p. 692). Guilford Press.

Moon K (2023). processR: Implementation of the 'PROCESS' Macro. R package version 0.2.8, commit c6c817b5b716f2615b9814cd5022aef1cbc5c09d, https://github.com/cardiomoon/processR.